HU17 endpoints are organized
