+++ src/main/java/com/pragma/powerup/usermicroservice/domain/usecase/OwnerUseCase.java
@@ -34,10 +34,7 @@ public class OwnerUseCase implements IOwnerServicePort {
 
         user.setRole(role);
         user.setPassword(authEncoderPort.encodePassword(user.getPassword()));
-        if (UserValid.isOlder(user)) {
-            userPersistencePort.saveUser(user);
-        } else {
-            throw new IsOlder();
-        }
+
+        userPersistencePort.saveUser(user);
     }
 }
